fix: corrigiendo busquedas

+++ currencyExchange/model/currencyExchange/index.js
@@ -33,9 +33,9 @@ class CurrencyExchangeRequest {
         }
     }
 
-    static async getExchangeRequestDetails(requestId) {
+    static async getExchangeRequestDetails(requestId, userId) {
         try {
-            const request = await this.findById(requestId);
+            const request = await this.findOne({ _id: requestId, id_usuario: userId });
             return request;
         } catch (error) {
             console.error("Error al listar la solicitud:", error);
@@ -43,9 +43,10 @@ class CurrencyExchangeRequest {
         }
     }
     
-    static async deleteExchangeRequest(requestId) {
+    
+    static async deleteExchangeRequest(requestId, userId) {
         try {
-            const result = await this.deleteOne({ _id: requestId });
+            const result = await this.deleteOne({ _id: requestId, id_usuario: userId});
             return result;
         } catch (error) {
             console.error("Error al eliminar", error);
